[FEATURE] File download checks date and etag
File download requests will now track Date and ETag headers. These will be used in If-Modified-Since (Date) and If-None-Match (ETag) headers in subsequent requests. This will allow the remote server to issue a 304 Not Modified response that will cause the PersistFileDataEventHandler to stop processing the file data.

This feature also introduces the `metadata` field in the remote ID mapping table and related methods.