[FEATURE] Override the field type for relations
In UserTS you can override the TCA type of the field. E.g.: change "text" to "inline". The value is always set to the field's current type. It is passed through the stdWrap `tx_interest.relationOverrides.<table>.<field>` and can be modified. The returned value must be a valid relation field type.