[BUGFIX] Correct sorting on foreign side of relations
Records with MM relations now store information about correct relation sorting in the mapping table's meta data.

This is used when a foreign record is a pending relation. When the foreign relation is inserted the local relation field is updated with the correct order as well.

This is also used when a foreign relation is updated, in case the relation is missing or the order is wrong on the other side.