[BUGFIX] Correct media file names (YouTube, etc.)
The file name will now always have the correct extension added. If the sys_file table's `name` property is left empty in the request, the name will be fetched from the media source. If the name is still empty and exception will be thrown. If the name parameter is used, this will override the file name (extension will be added if necessary).