[BUGFIX] Reattempt cache clear after DeadlockException
In some cases a database deadlock will occur when the DataHandler is clearing caches. This rarely happens when executing manual updates, but can happen from time to time when large amounts of data are updated, such as in this extension.

The DeadlockException thrown by Doctrine is now cought and clearing cache reattempted.

Resolves: #69