[!!!][FEATURE] Use DTO for record representation and identity (#77)
This internal change does not affect any external APIs, but improves the way record data and identities are handled. This fixes and issue that made multilingual operations impossible, as well as making it easier to work with record data and identities.

* `\Pixelant\Interest\Domain\Model\Dto\RecordRepresentation` represents a record state, and includes data as well as a `RecordInstanceIdentifier` object.
* `\Pixelant\Interest\Domain\Model\Dto\RecordInstanceIdentifier` represents a record's identity: table, UID, language, workspace, and remote ID. It can output both the remote ID as it is seen from the outside (which is only unique when combined with language and workspace) and the internal, so-called "aspected", remote ID, which is unique for each language and workspace combination.

Also fixes the condition for group relations. Was only accepting `internal_type` as "db", but `allowed` can also indicate a database relation.

Co-authored-by: Mats Svensson <mats@pixelant.se>