[FEATURE] HttpResponseEvent for post-processing responses (#79)
Introduces `HttpResponseEvent` which can be used to modify or process the response before it is sent to TYPO3. The event constructor takes one argument, a `ResponseInterface`-compatible object.

Also fixes the implementation of data or metaData as JSON string. If data or metaData is a string, we try to decode it as JSON. In some cases it can be easier to supply JSON within JSON rather than just supplying one big JSON array.